webapp/pyproject.toml
[tool.poetry]
name = "webapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
llama-index = "^0.10.59"
pandas = "^2.2.2"
llama-index-llms-gemini = "^0.1.11"
llama-index-embeddings-gemini = "^0.1.8"
beautifulsoup4 = "^4.12.3"
jinja2 = "^3.1.4"

[tool.poetry.group.dev.dependencies]
ruff = "*"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
